refactor csv to re-use
